+++ HeartyMonitorUITests/HeartyMonitor WatchKit Extension/InterfaceController.swift
@@ -170,7 +170,8 @@ class InterfaceController: WKInterfaceController, HKWorkoutSessionDelegate {
         guard let heartRateSamples = samples as? [HKQuantitySample] else {return}
         
         dispatch_async(dispatch_get_main_queue()) {
-            guard let sample = heartRateSamples.first else{return}
+            guard self.workoutActive else{return}
+            guard let sample = heartRateSamples.last else{return}
             let value = sample.quantity.doubleValueForUnit(self.heartRateUnit)
             guard value >= 0 && value <= 300 else{return}
             self.label.setText(String(UInt16(value)))